Remove take(1) from oneOf
Because it prevents multiple event on the trigger. For instance if the
user denies a request and ask again for it (via the trigger), it won't
work.

The caveat is request method don't work anymore, because of the toList
which requires a onComplete to works properly. The next fix will concern
that.